[Refactor] todo list에 context API 적용
