+++ vite-project/src/App.jsx
@@ -1,44 +1,21 @@
-import { useState } from 'react';
+import { useContext, useState } from 'react';
 import './App.css';
 import Button from './Button';
 import Input from './Input';
+import { TodoContext } from './context/TodoContext';
 
 function App() {
-  const [todos, setTodos] = useState([
-    { id: 1, task: '투두 만들어보기' },
-    { id: 2, task: '효진 민지 채은 희원 은서' },
-  ]);
-
-  const [text, setText] = useState('');
-  const [editingId, setEditingId] = useState('');
-  const [editText, setEditText] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-  };
-
-  const addTodo = () => {
-    if (text.trim().length === 0) {
-      alert('입력해라!');
-      return;
-    }
-    setTodos((prev) => [
-      ...prev,
-      { id: Math.floor(Math.random() * 100) + 2, task: text },
-    ]);
-    setText('');
-  };
-
-  const deleteTodo = (id) => {
-    setTodos((prev) => prev.filter((item) => item.id !== id));
-  };
-
-  const updateTodo = (id, text) => {
-    setTodos((prev) =>
-      prev.map((item) => (item.id === id ? { ...item, task: text } : item))
-    );
-    setEditingId('');
-  };
+  const {todos,
+    text,
+    setText,
+    editText,
+    editingId,
+    setEditingId,
+    setEditText,
+    handleSubmit,
+    addTodo,
+    deleteTodo,
+    updateTodo,} = useContext(TodoContext);
 
   return (
     <div className="container">
