+++ vite-project/src/main.jsx
@@ -2,9 +2,10 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
+import { TodoContext, TodoContextProvider } from './context/TodoContext.jsx'
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
+  <TodoContextProvider>
+    <App/>
+  </TodoContextProvider>
 )
